MountService: Add support for multiple storage configurations
Currently AOSP only supports a single storage configuration that is hardcoded
in storage_list.xml. Some device updates are now merging data and internal
storage partitions, and as a result require the storage list to support two
different variations.

By adapting code from the EriManager the MountService is able to load a custom
storage list from /data/storage_list.xml. If this file does not exist it will
default to the hardcoded values as before.

Change-Id: Id795520425ea3e99a5234e10c978dcdbd68a1311